VaidyaMitra-main/backend/requirements.txt
# Core Framework
fastapi==0.115.12
uvicorn[standard]==0.34.3
python-multipart==0.0.20
pydantic>=2.5.0
pydantic-settings>=2.1.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# PDF Processing
PyPDF2>=3.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Image Processing
Pillow>=10.0.0
pytesseract>=0.3.10

# Utilities
python-dotenv>=1.0.0
mangum==0.21.0
opencv-python-headless>=4.9.0
pdf2image>=1.17.0
PyMuPDF>=1.23.0
